Update Discord-RPC headers and move discord related code to a subdirectory.
