+++ xlive/discord/DiscordInterface.h
