+++ xlive/discord/discord-rpc-debug.lib
