+++ xlive/discord/discord-rpc.lib
