+++ xlive/dllmain.cpp
@@ -4,7 +4,6 @@
 #include "H2MOD.h"
 #include <iostream>
 #include <Shellapi.h>
-#include "discord-rpc.h"
 #include "H2Startup.h"
 #include "H2OnscreenDebugLog.h"
 #include "GSRunLoop.h"
