+++ xlive/xliveless.cpp
@@ -7,7 +7,7 @@
 #include <codecvt>
 #include <unordered_map>
 #include "Globals.h"
-#include "DiscordInterface.h"
+#include "discord/DiscordInterface.h"
 #include "H2Config.h"
 
 using namespace std;
